Move PlayerSummary to its own module
